requirements.txt
# Core Dependencies
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
pydantic==1.8.2

# AI/ML
langchain==0.0.335
openai==0.27.0
chromadb==0.4.15
numpy==1.24.3
pandas==2.1.3

# Blockchain & Crypto
web3==6.11.1
moralis==0.1.34
python-binance==1.0.19

# IoT & Monitoring
paho-mqtt==1.6.1
azure-iot-device==1.30.1
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0

# Database
sqlalchemy==1.4.23
psycopg2-binary==2.9.9
redis==4.3.4
motor==3.3.1
aiosqlite==0.17.0

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.5

# Testing
pytest==6.2.5
pytest-asyncio==0.21.1
httpx==0.23.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1 